% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeStarOutput.R
\name{mergeStarOutput}
\alias{mergeStarOutput}
\title{merge multiple files with the more than one intersected rownames}
\usage{
mergeStarOutput(
  pattern = ".*count.tab",
  header = F,
  sep = "\\t",
  stringsAsFactors = F,
  namesSep = "\\\\.",
  colSelect = c(1, 2),
  nameSelect = 1
)
}
\arguments{
\item{pattern}{character string containing a regular expression to be matched in the given character vector.}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line.}

\item{sep}{the field separator character of the separate file.}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}

\item{namesSep}{colnames separator.}

\item{colSelect}{column index to merge.}

\item{nameSelect}{name index of separated colnames to be new colnames.}
}
\value{
generate a expression matrix (eSet).
}
\description{
merge multiple files with the more than one intersected rownames
}
\examples{
 mergeStarOutput(pattern =  ".*count.tab",header = F,sep = "\t",stringsAsFactors = F,namesSep="\\\\.",colSelect=c(1,2),nameSelect=1)
}
\keyword{downGSE}
